+++ flutter_inappwebview/lib/src/in_app_webview/headless_in_app_webview.dart
@@ -364,8 +364,14 @@ class HeadlessInAppWebView {
       PermissionRequest permissionRequest,
     )?
     onPermissionRequest,
+    @Deprecated('Use onFaviconChanged instead')
     void Function(InAppWebViewController controller, Uint8List icon)?
     onReceivedIcon,
+    void Function(
+      InAppWebViewController controller,
+      FaviconChangedRequest faviconChangedRequest,
+    )?
+    onFaviconChanged,
     void Function(InAppWebViewController controller, LoginRequest loginRequest)?
     onReceivedLoginRequest,
     void Function(
@@ -436,6 +442,26 @@ class HeadlessInAppWebView {
       ProcessFailedDetail detail,
     )?
     onProcessFailed,
+    FutureOr<NotificationReceivedResponse?> Function(
+      InAppWebViewController controller,
+      NotificationReceivedRequest request,
+    )?
+    onNotificationReceived,
+    FutureOr<SaveAsUIShowingResponse?> Function(
+      InAppWebViewController controller,
+      SaveAsUIShowingRequest request,
+    )?
+    onSaveAsUIShowing,
+    FutureOr<SaveFileSecurityCheckStartingResponse?> Function(
+      InAppWebViewController controller,
+      SaveFileSecurityCheckStartingRequest request,
+    )?
+    onSaveFileSecurityCheckStarting,
+    FutureOr<ScreenCaptureStartingResponse?> Function(
+      InAppWebViewController controller,
+      ScreenCaptureStartingRequest request,
+    )?
+    onScreenCaptureStarting,
     void Function(
       InAppWebViewController controller,
       AcceleratorKeyPressedDetail detail,
@@ -446,6 +472,11 @@ class HeadlessInAppWebView {
       ShowFileChooserRequest request,
     )?
     onShowFileChooser,
+    FutureOr<LaunchingExternalUriSchemeResponse?> Function(
+      InAppWebViewController controller,
+      LaunchingExternalUriSchemeRequest request,
+    )?
+    onLaunchingExternalUriScheme,
   }) : this.fromPlatformCreationParams(
          params: PlatformHeadlessInAppWebViewCreationParams(
            controllerFromPlatform:
@@ -726,6 +757,10 @@ class HeadlessInAppWebView {
            onReceivedIcon: onReceivedIcon != null
                ? (controller, icon) => onReceivedIcon.call(controller, icon)
                : null,
+             onFaviconChanged: onFaviconChanged != null
+               ? (controller, request) =>
+                 onFaviconChanged.call(controller, request)
+               : null,
            androidOnReceivedTouchIconUrl: androidOnReceivedTouchIconUrl != null
                ? (controller, url, precomposed) => androidOnReceivedTouchIconUrl
                      .call(controller, url, precomposed)
@@ -822,6 +857,23 @@ class HeadlessInAppWebView {
                ? (controller, detail) =>
                      onProcessFailed.call(controller, detail)
                : null,
+               onNotificationReceived: onNotificationReceived != null
+               ? (controller, request) =>
+                 onNotificationReceived.call(controller, request)
+               : null,
+               onSaveAsUIShowing: onSaveAsUIShowing != null
+               ? (controller, request) =>
+                 onSaveAsUIShowing.call(controller, request)
+               : null,
+               onSaveFileSecurityCheckStarting:
+               onSaveFileSecurityCheckStarting != null
+               ? (controller, request) =>
+                 onSaveFileSecurityCheckStarting.call(controller, request)
+               : null,
+               onScreenCaptureStarting: onScreenCaptureStarting != null
+               ? (controller, request) =>
+                 onScreenCaptureStarting.call(controller, request)
+               : null,
            onAcceleratorKeyPressed: onAcceleratorKeyPressed != null
                ? (controller, detail) =>
                      onAcceleratorKeyPressed.call(controller, detail)
@@ -829,6 +881,10 @@ class HeadlessInAppWebView {
            onShowFileChooser: onShowFileChooser != null
                ? (controller, request) =>
                      onShowFileChooser.call(controller, request)
+               : null,
+           onLaunchingExternalUriScheme: onLaunchingExternalUriScheme != null
+               ? (controller, request) =>
+                     onLaunchingExternalUriScheme.call(controller, request)
                : null,
          ),
        );